+++ app/controllers/edge/api/edge_configuration_controller.rb
@@ -21,21 +21,6 @@ class EdgeConfigurationController < RestController
     end
   end
 
-  def validate_permission
-    log_message = "Validating permissions to run edge api for user '#{current_user.id}'"
-    logger.debug(LogMessages::Endpoints::EndpointRequested.new(log_message))
-    allowed_params = %i[account]
-    options = params.permit(*allowed_params).to_h.symbolize_keys
-    begin
-      validate_conjur_admin_group(options[:account])
-      head(200)
-      logger.debug(LogMessages::Endpoints::EndpointFinishedSuccessfully.new(log_message))
-    rescue => e
-      logger.error(LogMessages::Conjur::GeneralError.new(e.message))
-      raise e
-    end
-  end
-
   def get_role
     log_message = "Validating role '#{current_user.id}'"
     logger.debug(LogMessages::Endpoints::EndpointRequested.new(log_message))
@@ -52,4 +37,39 @@ class EdgeConfigurationController < RestController
       raise e
     end
   end
+
+  EDGE_NOT_FOUND = "Edge not found"
+  def get_edge_info
+    logger.debug(LogMessages::Endpoints::EndpointRequested.new("GET /agents/#{params[:account]}/#{params[:identifier]}/info"))
+    allowed_params = %i[account identifier]
+    options = params.permit(*allowed_params).to_h.symbolize_keys
+    verify_edge_host(options)
+    if params[:identifier] != Edge.hostname_to_id(current_user.role_id)
+      logger.error(
+        Errors::Authorization::EndpointNotVisibleToRole.new(
+          "Requested identifier #{params[:identifier]} is not allowed for user #{current_user.role_id}"
+        )
+      )
+      raise ApplicationController::Forbidden
+    end
+    begin
+      edge = Edge.where(id: params[:identifier]).first
+      if edge
+        render(json: {
+          id: edge.id,
+          name: edge.name
+        })
+      else
+        raise Exceptions::RecordNotFound.new(params[:identifier], message: EDGE_NOT_FOUND)
+      end
+      logger.debug(LogMessages::Endpoints::EndpointFinishedSuccessfully.new("GET /agents/#{params[:account]}/#{params[:identifier]}/info"))
+    rescue Exceptions::RecordNotFound => e
+      @error_message = e.message
+      raise Exceptions::RecordNotFound.new(params[:identifier], message: EDGE_NOT_FOUND)
+    rescue => e
+      logger.error(LogMessages::Conjur::GeneralError.new(e.message))
+      raise e
+    end
+  end
 end
+